Add Github Actions
Github Actions will run and build all artifacts when a pull request is
made or a branch is pushed. There is a cache that stores all bazel
binaries (bazel, bazelisk, etc) and the bazel outputs themselves, so
that caching between runs is preserved.

Dependabot also updates the github actions version every week.