[autofresh] Add recursive watcher
Start port of autofresh to monorepo by adding a recursive watcher that
recursively watches for changes in a directory and all of its
subdirectories. It wraps fsnotify, because fsnotify does not recursively
watch directories currently.

Tests are included; tests might be slightly flaky due to file system
dependencies. Solution is to bump up the sleep times to wait for
recursive watcher to catch up.