Split PR and master CI into different jobs
By splitting job into two different jobs, we can differentiate the
master cache and the PR cache. The PR will fallback to the master cache
on the first run when the PR cache is not yet populated.

This PR also updates the README, so if the bazel cache is properly
accessed, then the build and test portion of this CI workflow should be
fast.