Setup Hermetic Python and pip
Previously, we tried out the Nix setup for Bazel Python, but found that
Nix largely relies on the system version of Python, whereas we would
like a separate version of Python and pip.

This commit ensures that we use the `@rules_python` setup for standalone
Python, which is mostly hermetic (and much faster to download instead of
building Python from scratch). Additionally, we also setup the lazily
fetched `pip_parse` setup, with each "project" capable of having its own
pip repository.