Look into Docker support
This adds proper Docker support configuration to the WORKSPACE toolchain
definition.

Additionally, container_push rules are implemented and a CI pipeline to
push an image to the Github Container Repository is created. A new image
will be pushed to Github, tagged with the commit hash. The commit hash
is acquired from Bazel via the workspace status command, with the
`--stamp` flag or the stamp always rule flag.

A `multirun` rule is added in the root BUILD.bazel file to push multiple
images. A CI job is added to the master GitHub action that runs after
the master test passes, which logs into the ghcr.io container registry
in Docker and runs this rule to push all images into the registry.