Fastlinks Local File
Add support for the local file storage into fastlinks. There is now a
configuration option to switch between the database storage and the
local file bbolt storage. By default it will use the local storage.

A new utility function to "touch" a file was added into utils-go. The
intent is to create a file if it did not yet exist.

Because Bazel strips environment variables during testing, there, we
must export the `$HOME` variable over to the tests, so that. This should
be safe env var because the `$HOME` env var should not change very often
(usually never) between runs and is cache friendly.