[fastlinks] Add static template HTML generation and controllers
Add setup for template generation to fastlinks. Create a new view
struct that takes a sample template, and produces the correct output.

Templates are embeded into the go binary with the directive
`//go:embed`, and we use the `embedsrc` field in the `go_library` Bazel
rule to embed the HTML templates into the binary statically.

Add controllers to interact with the stores to create, get and delete
routes.

Connected adapters to controllers, implementing basic functionality for
fastlinks golinks.